refactor: split in-memory testing kit modules
